• Web Service that accepts search criteria to build and return matching trips.  Ex: A one way trip from Montreal to Vancouver.  Ex: A roundtrip between Montreal and Vancouver • Make the API RESTful • Cache search results • Calculate total distance traveled in a trip • Sort trip listing (by total cost, duration, etc.) • Filter trips (by airline, total duration, total cost, etc.)

+++ database/migrations/2023_05_18_034205_create_airports_table.php
@@ -19,8 +19,8 @@ class CreateAirportsTable extends Migration
             $table->string('name');
             $table->string('city');
             $table->string('country_code');
-            $table->float('latitude');
-            $table->float('longitude');
+            $table->double('latitude');
+            $table->double('longitude');
             $table->string('timezone');
             $table->timestamps();
         });
